m_fix : revive death charactor btw quotes
